pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial_analyzer"
version = "0.1.0"
description = "Production-grade financial analysis pipeline with yfinance, Pydantic, SQLAlchemy, and Typer CLI"
authors = [{name = "Gulam Mazid", email = "mazidgulam786@gmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0.0",
  "yfinance>=0.2.40",
  "pydantic>=2.7.0",
  "typer[all]>=0.12.0",
  "SQLAlchemy>=2.0.0",
  "PyYAML>=6.0.0",
  "numpy>=1.24.0"
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
